20172689_reporte2.md
What is the research question of the article?
In this article on econometric models, the objective or research question is, what estimation and inference methods will be used for a High-Dimensional Sparse (HDS) econometric model; that is, when there are a large number of regressors.

What are the strengths and weaknesses of the document's approach to answering that question?
One of the strengths is that it begins to explain it with simple methods and practical examples to understand it better. Then he digs deep and uses slightly more complex methods like Lasso and others. Subsequently, a comparison of the methods is made, in which various problems arise that seek to be solved with various instruments such as the Double Loop. Accompanied by examples, it is what makes the article a little more digestible; However, one difficulty would be that if you don't have intermediate econometrics skills, it would be difficult to read. 

How does this document advance knowledge about the question, that is, what is the contribution?
As before, the great variety of methods, of different articles, for HDS models. In this way, it aims to find the best regressors for a more rigorous estimation. In addition, it shows how empirically, with methodological improvements, the regressors can be improved to make better estimates on issues such as educational policy and life in general.

What would be one or two valuable and specific next steps to advance this question?
As mentioned at the end of the article, there are many avenues for additional research. The use of ℓ1 penalization is only one method of performing estimation with high-dimensional data (HDS). In addition to this limitation, it would be interesting to update the article and apply it not only to developed countries such as the USA.